+++ presentation/src/main/res/layout/activity_my_todo_create.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="vm"
-            type="com.going.presentation.todo.mytodo.create.MyTodoCreateViewModel" />
+            type="com.going.presentation.todo.mytodo.detail.MyTodoCreateViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
